backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-dotenv
pillow
torch
torchvision

faiss-cpu
numpy